# fix: slow computation on type conversion