--- a/pkg/accounting/pnl/report.go
+++ b/pkg/accounting/pnl/report.go
@@ -14,7 +14,7 @@ import (
 )
 
 type AverageCostPnlReport struct {
-	LastPrice float64      `json:"lastPrice"`
+	LastPrice fixedpoint.Value      `json:"lastPrice"`
 	StartTime time.Time    `json:"startTime"`
 	Symbol    string       `json:"symbol"`
 	Market    types.Market `json:"market"`
@@ -23,12 +23,12 @@ type AverageCostPnlReport struct {
 	Profit           fixedpoint.Value   `json:"profit"`
 	NetProfit        fixedpoint.Value   `json:"netProfit"`
 	UnrealizedProfit fixedpoint.Value   `json:"unrealizedProfit"`
-	AverageCost      float64            `json:"averageCost"`
-	BuyVolume        float64            `json:"buyVolume,omitempty"`
-	SellVolume       float64            `json:"sellVolume,omitempty"`
-	FeeInUSD         float64            `json:"feeInUSD"`
-	Stock            float64            `json:"stock"`
-	CurrencyFees     map[string]float64 `json:"currencyFees"`
+	AverageCost      fixedpoint.Value            `json:"averageCost"`
+	BuyVolume        fixedpoint.Value            `json:"buyVolume,omitempty"`
+	SellVolume       fixedpoint.Value            `json:"sellVolume,omitempty"`
+	FeeInUSD         fixedpoint.Value            `json:"feeInUSD"`
+	Stock            fixedpoint.Value            `json:"stock"`
+	CurrencyFees     map[string]fixedpoint.Value `json:"currencyFees"`
 }
 
 func (report *AverageCostPnlReport) JSON() ([]byte, error) {
@@ -38,26 +38,26 @@ func (report *AverageCostPnlReport) JSON() ([]byte, error) {
 func (report AverageCostPnlReport) Print() {
 	log.Infof("TRADES SINCE: %v", report.StartTime)
 	log.Infof("NUMBER OF TRADES: %d", report.NumTrades)
-	log.Infof("AVERAGE COST: %s", types.USD.FormatMoneyFloat64(report.AverageCost))
-	log.Infof("TOTAL BUY VOLUME: %f", report.BuyVolume)
-	log.Infof("TOTAL SELL VOLUME: %f", report.SellVolume)
-	log.Infof("STOCK: %f", report.Stock)
+	log.Infof("AVERAGE COST: %s", types.USD.FormatMoney(report.AverageCost))
+	log.Infof("TOTAL BUY VOLUME: %s", report.BuyVolume.String())
+	log.Infof("TOTAL SELL VOLUME: %s", report.SellVolume.String())
+	log.Infof("STOCK: %s", report.Stock.String())
 
 	// FIXME:
 	// log.Infof("FEE (USD): %f", report.FeeInUSD)
-	log.Infof("CURRENT PRICE: %s", types.USD.FormatMoneyFloat64(report.LastPrice))
+	log.Infof("CURRENT PRICE: %s", types.USD.FormatMoney(report.LastPrice))
 	log.Infof("CURRENCY FEES:")
 	for currency, fee := range report.CurrencyFees {
-		log.Infof(" - %s: %f", currency, fee)
+		log.Infof(" - %s: %s", currency, fee.String())
 	}
-	log.Infof("PROFIT: %s", types.USD.FormatMoneyFloat64(report.Profit.Float64()))
-	log.Infof("UNREALIZED PROFIT: %s", types.USD.FormatMoneyFloat64(report.UnrealizedProfit.Float64()))
+	log.Infof("PROFIT: %s", types.USD.FormatMoney(report.Profit))
+	log.Infof("UNREALIZED PROFIT: %s", types.USD.FormatMoney(report.UnrealizedProfit))
 }
 
 func (report AverageCostPnlReport) SlackAttachment() slack.Attachment {
 	var color = slackstyle.Red
 
-	if report.UnrealizedProfit > 0 {
+	if report.UnrealizedProfit.Sign() > 0 {
 		color = slackstyle.Green
 	}
 
@@ -70,12 +70,12 @@ func (report AverageCostPnlReport) SlackAttachment() slack.Attachment {
 		Fields: []slack.AttachmentField{
 			{Title: "Profit", Value: types.USD.FormatMoney(report.Profit)},
 			{Title: "Unrealized Profit", Value: types.USD.FormatMoney(report.UnrealizedProfit)},
-			{Title: "Current Price", Value: report.Market.FormatPrice(report.LastPrice), Short: true},
-			{Title: "Average Cost", Value: report.Market.FormatPrice(report.AverageCost), Short: true},
+			{Title: "Current Price", Value: report.Market.FormatPrice(report.LastPrice.Float64()), Short: true},
+			{Title: "Average Cost", Value: report.Market.FormatPrice(report.AverageCost.Float64()), Short: true},
 
 			// FIXME:
 			// {Title: "Fee (USD)", Value: types.USD.FormatMoney(report.FeeInUSD), Short: true},
-			{Title: "Stock", Value: strconv.FormatFloat(report.Stock, 'f', 8, 64), Short: true},
+			{Title: "Stock", Value: report.Stock.String(), Short: true},
 			{Title: "Number of Trades", Value: strconv.Itoa(report.NumTrades), Short: true},
 		},
 		Footer:     report.StartTime.Format(time.RFC822),