--- a/pkg/bbgo/config.go
+++ b/pkg/bbgo/config.go
@@ -124,7 +124,7 @@ func (m BacktestAccountBalanceMap) BalanceMap() types.BalanceMap {
 		balances[currency] = types.Balance{
 			Currency:  currency,
 			Available: value,
-			Locked:    0,
+			Locked:    fixedpoint.Zero,
 		}
 	}
 	return balances