--- a/pkg/bbgo/context.go
+++ b/pkg/bbgo/context.go
@@ -8,4 +8,3 @@ import (
 type Context struct {
 	sync.Mutex
 }
-