before:
```bash
goos: linux
goarch: amd64
pkg: github.com/c9s/bbgo/pkg/fixedpoint
cpu: Intel(R) Core(TM) i7-9700 CPU @ 3.00GHz
BenchmarkMul/mul-float64-8         	281725761	         4.052 ns/op
BenchmarkMul/mul-float64-large-numbers-8         	295906885	         4.049 ns/op
BenchmarkMul/mul-big-small-numbers-8             	19299628	       104.5 ns/op
BenchmarkMul/mul-big-large-numbers-8             	19619779	        92.10 ns/op
PASS
ok  	github.com/c9s/bbgo/pkg/fixedpoint	7.144s
```

after:
```bash
goos: linux
goarch: amd64
pkg: github.com/c9s/bbgo/pkg/fixedpoint
cpu: Intel(R) Core(TM) i7-9700 CPU @ 3.00GHz
BenchmarkMul/mul-float64-8         	1000000000	         0.2381 ns/op
BenchmarkMul/mul-float64-large-numbers-8         	1000000000	         0.2401 ns/op
BenchmarkMul/mul-big-small-numbers-8             	18941152	        74.09 ns/op
BenchmarkMul/mul-big-large-numbers-8             	16903243	        99.72 ns/op
PASS
ok  	github.com/c9s/bbgo/pkg/fixedpoint	3.757s
```