@zenixls2 2022-01-27
also changed  the behavior of floatpoint presentation.  
use 'g' instead of 'f' to present price for BTC pairs.