@c9s 2022-01-27
it's faster because it now uses float64 instead of int64?