@zenixls2 2022-01-28
update: change float64 in types/kline to fixedpoint.Value