@zenixls2 2022-02-07
based on https://github.com/apmckinlay/gsuneido/blob/master/util/dnum/dnum.go
with some minor optimization
```
goos: linux
goarch: amd64
pkg: github.com/c9s/bbgo/pkg/fixedpoint
cpu: Intel(R) Core(TM) i7-9700 CPU @ 3.00GHz
BenchmarkMul/mul-float64-8         	69861969	        17.11 ns/op
BenchmarkMul/mul-float64-large-numbers-8         	16960190	        70.20 ns/op
BenchmarkMul/mul-big-small-numbers-8             	14283976	       106.1 ns/op
BenchmarkMul/mul-big-large-numbers-8             	12499324	       111.6 ns/op
PASS
ok  	github.com/c9s/bbgo/pkg/fixedpoint	5.576s
```