@zenixls2 2022-02-15
changes:
- add dnum implementation (to build, set -tags=dnum) (without dnum flag, use legacy convert.go implementation)
- add more tests, fix direct operations on fixedpoint.Value. 
- fix accounting bug in backtest (startPrice => lastPrice in finalQuoteAsset)
- add skeleton as a buy-and-hold strategy