@zenixls2 2022-02-15
This pr also resolves #240, #396 and #400   
for #400,  example:  
```bash
go run -tags dnum ./cmd/bbgo backtest --exchange binance --config config/abc.yaml -v 0
```
only workable when using dnum tag.